+++ src/core/vector/qgsvectorlayerfeaturecounter.h
@@ -56,13 +56,13 @@ class CORE_EXPORT QgsVectorLayerFeatureCounter : public QgsTask
      *
      * \note Not available in Python bindings.
      */
-    QHash<QString, long> symbolFeatureCountMap() const SIP_SKIP;
+    QHash<QString, long long> symbolFeatureCountMap() const SIP_SKIP;
 
     /**
      * Returns the feature count for a particular \a legendKey.
      * If the key has not been found, -1 will be returned.
      */
-    long featureCount( const QString &legendKey ) const;
+    long long featureCount( const QString &legendKey ) const;
 
     /**
      * Returns the QgsFeatureIds for each symbol. Only valid after the symbolsCounted()
@@ -93,7 +93,7 @@ class CORE_EXPORT QgsVectorLayerFeatureCounter : public QgsTask
     std::unique_ptr<QgsVectorLayerFeatureSource> mSource;
     std::unique_ptr<QgsFeatureRenderer> mRenderer;
     QgsExpressionContext mExpressionContext;
-    QHash<QString, long> mSymbolFeatureCountMap;
+    QHash<QString, long long> mSymbolFeatureCountMap;
     QHash<QString, QgsFeatureIds> mSymbolFeatureIdMap;
     bool mWithFids = false;
     int mFeatureCount;
